Removed TIR scaling in thermal sharpening function
The sharpening function should work the same independent of the scaling of the TIR band.

Renamed the sharpening function to "landsat", since the thermal and sharpening are in the module names and this function is Landsat specific.

Added tests for scaled and unscaled SR images and TOA images.

Removed the integer casting in the function (this will need to be added back into the export utility to return SR like TIR images.